feat(AXD-27): refactor sanity queries for posts, projects and tags
