--- src/clients/sanity/converters/convertSanityPostToPost/convertSanityPostToPost.interface.ts
@@ -1,12 +1,0 @@
-import type { SanityPostsGroqQueryType } from 'src/clients/sanity/queries'
-import type { ConvertSanityDocumentToTypeFunction, PostType } from 'src/types'
-
-export interface ConvertSanityPostToPostFunctionRetrieveFieldsI {
-  body?: boolean
-}
-
-export type ConvertSanityPostToPostFunction = ConvertSanityDocumentToTypeFunction<
-  SanityPostsGroqQueryType,
-  PostType,
-  ConvertSanityPostToPostFunctionRetrieveFieldsI
->
