--- src/clients/sanity/converters/convertSanityPostToPost/index.ts
@@ -1,1 +1,0 @@
-export { default as convertSanityPostToPost } from './convertSanityPostToPost'
