--- src/clients/sanity/converters/convertSanityProjectToProject/convertSanityProjectToProject.interface.ts
@@ -1,5 +1,0 @@
-import type { SanityProjectsGroqQueryType } from 'src/clients/sanity/queries'
-
-import type { ConvertSanityDocumentToTypeFunction, ProjectType } from 'src/types'
-
-export type ConvertSanityProjectToProjectFunction = ConvertSanityDocumentToTypeFunction<SanityProjectsGroqQueryType, ProjectType>
