--- src/clients/sanity/converters/convertSanityProjectToProject/index.ts
@@ -1,1 +1,0 @@
-export { default as convertSanityProjectToProject } from './convertSanityProjectToProject'
