--- src/clients/sanity/converters/convertSanityTagToTag/convertSanityTagToTag.interface.ts
@@ -1,4 +1,0 @@
-import type { SanityTagsGroqQueryType } from 'src/clients/sanity/queries'
-import type { ConvertSanityDocumentToTypeFunction, TagType } from 'src/types'
-
-export type ConvertSanityTagToTagFunction = ConvertSanityDocumentToTypeFunction<SanityTagsGroqQueryType, TagType>
