--- src/clients/sanity/converters/convertSanityTagToTag/index.ts
@@ -1,1 +1,0 @@
-export { default as convertSanityTagToTag } from './convertSanityTagToTag'
