--- src/clients/sanity/converters/index.ts
@@ -1,3 +1,0 @@
-export * from './convertSanityPostToPost'
-export * from './convertSanityTagToTag'
-export * from './convertSanityProjectToProject'
