+++ src/clients/sanity/index.ts
@@ -1,3 +1,2 @@
 export * from './getImageUrl'
 export * from './queries'
-export * from './converters'
