+++ src/clients/sanity/queries/index.ts
@@ -1,3 +1,3 @@
-export * from './Posts'
-export * from './Tags'
-export * from './Projects'
+export * from './posts'
+export * from './projects'
+export * from './tags'
