+++ src/clients/sanity/queries/posts/Posts.consts.ts
@@ -1,4 +1,4 @@
-import type { ConvertSanityPostToPostFunctionRetrieveFieldsI } from './convertSanityPostToPost.interface'
+import type { ConvertSanityPostToPostFunctionRetrieveFieldsI } from './Posts.interface'
 
 export const RETRIEVE_FIELDS_DEFAULTS: ConvertSanityPostToPostFunctionRetrieveFieldsI = {
   body: false,
