+++ src/clients/sanity/queries/posts/index.ts
@@ -1,0 +1,1 @@
+export { default as getPosts } from './Posts.query'
