+++ src/clients/sanity/queries/projects/index.ts
@@ -1,0 +1,2 @@
+export { default as getProjects } from './Projects.query'
+export type { ProjectType } from './Projects.interface'
