+++ src/clients/sanity/queries/tags/Tags.query.ts
@@ -1,0 +1,12 @@
+import { useSanityClient } from '@sanity/astro'
+
+import convertSanityTagToTag from './Tags.converter'
+import { sanityTagsGroqQuery } from './Tags.groq'
+import type { SanityTagType } from './Tags.interface'
+
+const getTags = async (prefix?: string | undefined | boolean | null) => {
+  const query = await useSanityClient().fetch<SanityTagType[]>(sanityTagsGroqQuery)
+  return convertSanityTagToTag(query, { prefix })
+}
+
+export default getTags
