+++ src/clients/sanity/queries/tags/index.ts
@@ -1,0 +1,1 @@
+export { default as getTags } from './Tags.query'
