+++ src/layouts/BlogPostLayout.astro
@@ -6,7 +6,7 @@ type Props = {
   title: string
   description: string
   createdAt: Date
-  updatedAt: Date
+  updatedAt?: Date
   heroImage?: string
 }
