+++ src/pages/project/index.astro
@@ -1,17 +1,11 @@
 ---
 import { FormattedDate } from 'src/components/react'
 import FeaturedImage from 'src/components/sanity/FeaturedImage.astro'
-import { useSanityClient } from '@sanity/astro'
 import { SITE_DESCRIPTION, SITE_TITLE } from 'src/config/consts'
 import Layout from 'src/layouts/Layout.astro'
-import { type SanityProjectsGroqQueryType, convertSanityProjectsToProject, sanityProjectsGroqQuery } from 'src/clients/sanity'
+import { getProjects } from 'src/clients/sanity'
 
-export async function getBlogProjects() {
-  return convertSanityProjectsToProject(await useSanityClient().fetch<SanityProjectsGroqQueryType[]>(sanityProjectsGroqQuery), {
-    prefix: false,
-  })
-}
-const projects = await getBlogProjects()
+const projects = await getProjects(true)
 ---
 
 <Layout title={SITE_TITLE} description={SITE_DESCRIPTION}>
@@ -20,14 +14,14 @@ const projects = await getBlogProjects()
   </section>
   <section class="grid grid-cols-1 md:grid-cols-2 xl:grid-cols-3 xxl:grid-cols-4 grid-flow-row gap-4 md:gap-8 py-24">
     {
-      projects.map(({ featuredImage, title, createdAt, slug, estimatedReadingTime }) => (
+      projects.map(({ featuredImage, title, pubDate, slug, estimatedReadingTime }) => (
         <div class="flex flex-col">
           <div class="relative rounded-md overflow-hidden h-[208px]">
             {featuredImage && <FeaturedImage class="max-h-52" width={360} height={240} asset={featuredImage} alt={title} />}
           </div>
           <div class="flex flex-col md:flex-row text-primary mt-6">
             <div class="flex flex-col md:flex-row">
-              <FormattedDate date={createdAt} />
+              <FormattedDate date={pubDate} />
               <span class="px-2">•</span>
               <h4>Reading time: {estimatedReadingTime} min</h4>
             </div>
