+++ src/types/index.ts
@@ -1,5 +1,2 @@
-export type * from './PostType'
-export type * from './TagType'
 export type * from './ConvertSanityDocumentToTypeFunction'
-export type * from './ProjectType'
 export type * from './SanityAssetType'
