+++ tsconfig.json
@@ -1,6 +1,7 @@
 {
   "extends": "astro/tsconfigs/strictest",
   "compilerOptions": {
+    "exactOptionalPropertyTypes": false,
     "strictNullChecks": true,
     "jsx": "react-jsx",
     "jsxImportSource": "react",
